Implement address rescanning.
When a wallet is opened, a rescan request will be sent to btcd with
all active addresses from the wallet, to rescan from the last synced
block (now saved to the wallet file) and the current best block.

As multi-account support is further explored, rescan requests should
be batched together to send a single request for all addresses from
all wallets.

This change introduces several changes to the wallet, tx, and utxo
files.  Wallet files are still compatible, however, a rescan will try
to start at the genesis block since no correct "last synced to" or
"created at block X" was saved.  The tx and utxo files, however, are
not compatible and should be deleted (or an error will occur on read).
If any errors occur opening the utxo file, a rescan will start
beginning at the creation block saved in the wallet.

+++ tx/tx_test.go
@@ -144,7 +144,7 @@ func TestUtxoStoreWriteRead(t *testing.T) {
 		utxo.Out.Index = uint32(i + 2)
 		utxo.Subscript = []byte{}
 		utxo.Amt = uint64(i + 3)
-		utxo.Height = int64(i + 4)
+		utxo.Height = int32(i + 4)
 		*store1 = append(*store1, utxo)
 	}
